Issue #66:  Move INotificationService.refresh() to tagging interface IRefreshable, supporting both services and filters
