+++ notification-portlet-webapp/src/main/java/org/jasig/portlet/notice/service/AbstractNotificationService.java
@@ -89,9 +89,6 @@ public abstract class AbstractNotificationService implements INotificationServic
         return NotificationResponse.EMPTY_RESPONSE;
     }
 
-    @Override
-    public void refresh(HttpServletRequest request, HttpServletResponse response) { /* no-op */ }
-
     /*
      * Implementation
      */
